requirements.txt
marker-pdf
opencv-python
pypdf
python-dotenv
requests
anthropic
openai
rich